Preserve count precision in patch caches
Write offset-subtracted raw patches and clipped BM4D teachers as float32 instead of float16. Float16 increasingly coarsens count values across the uint16 sensor range, reaching multi-count quantization through the bright tail before the intensity transform or model sees an example. Float32 exactly represents the source integer counts while also retaining negative offset-subtracted samples and fractional BM4D estimates.

Remove the float16 saturation helper, use one shared count dtype for both memory-mapped arrays, and record count_dtype in each cache config for reproducibility. Update cache-layout documentation and extend the precompute configuration test to cover the persisted and allocated dtype.

Existing caches remain readable for backward compatibility but must be rebuilt to receive the precision improvement.

+++ scripts/visualize_patches.py
@@ -10,8 +10,8 @@ structures (noise, off-target label)?
 Works on either cache produced by scripts/precompute.py (train or val); both
 share the same layout::
 
-    raw.npy      float16  (N, *patch_shape)   offset-subtracted counts
-    teacher.npy  float16  (N, *patch_shape)   clipped BM4D denoising
+    raw.npy      float32  (N, *patch_shape)   offset-subtracted counts
+    teacher.npy  float32  (N, *patch_shape)   clipped BM4D denoising
     fg.npy       uint8    (N, *patch_shape)   foreground mask (0/1)
 
 Each patch is one row with five count-space panels:
